feat(langy): spine fold + message map join the shared package (ADR-059 Phase 4, server half)
foldLangyConversationState and mapLangyMessageEvent move to
@langwatch/langy as pure reducers — the server projections are now thin
rigs over shared logic for ALL THREE Langy projections (turn document,
conversation spine, message rows), completing the server side of the
one-projection-library story. No behaviour change; the existing fold and
map suites pin the delegation.

The spine state deliberately keeps the server-only columns (RunToken,
handoff tokens) — they are part of the fold's truth — with the module
doc stating they never enter any wire schema: the tail read stays
restricted to the turn vocabulary.

The client half of Phase 4 (the message list + spine folded in the
browser, retiring signal-then-refetch for the open conversation
entirely) deliberately waits for the Phase 3 slice to be validated live
— per the ADR's stop-and-review posture.
