+++ langwatch/src/server/app-layer/langy/repositories/__tests__/langy-conversation-projection.prisma.repository.unit.test.ts
@@ -8,7 +8,7 @@ import type { Event } from "~/server/event-sourcing/domain/types";
 import type { ProjectionStoreContext } from "~/server/event-sourcing/projections/projectionStoreContext";
 import type { StateProjectionDefinition } from "~/server/event-sourcing/projections/stateProjection.types";
 import { StateProjectionExecutor } from "~/server/event-sourcing/projections/stateProjectionExecutor";
-import type { LangyConversationStateData } from "~/server/event-sourcing/pipelines/langy-conversation-processing/projections/langyConversationState.foldProjection";
+import type { LangyConversationStateData } from "@langwatch/langy";
 import { PrismaLangyConversationProjectionRepository } from "../langy-conversation-projection.prisma.repository";
 
 type Row = LangyConversationProjection;
