+++ langwatch/src/server/app-layer/langy/repositories/__tests__/langy-message-projection.prisma.repository.unit.test.ts
@@ -2,7 +2,7 @@ import type { LangyMessageProjection } from "@prisma/client";
 import { describe, expect, it, vi } from "vitest";
 import { createTenantId } from "~/server/event-sourcing/domain/tenantId";
 import type { ProjectionStoreContext } from "~/server/event-sourcing/projections/projectionStoreContext";
-import type { LangyMessageProjectionRecord } from "~/server/event-sourcing/pipelines/langy-conversation-processing/projections/langyMessageOperational.mapProjection";
+import type { LangyMessageProjectionRecord } from "@langwatch/langy";
 import { PrismaLangyMessageProjectionRepository } from "../langy-message-projection.prisma.repository";
 
 type Row = LangyMessageProjection;
