+++ langwatch/src/server/event-sourcing/pipelineRegistry.ts
@@ -121,9 +121,9 @@ import type { SimulationRunStateRepository } from "./pipelines/simulation-proces
 import type { ComputeRunMetricsCommandData } from "./pipelines/simulation-processing/schemas/commands";
 import { SIMULATION_PROJECTION_VERSIONS } from "./pipelines/simulation-processing/schemas/constants";
 import { createLangyConversationProcessingPipeline } from "./pipelines/langy-conversation-processing/pipeline";
-import { type LangyConversationStateData } from "./pipelines/langy-conversation-processing/projections/langyConversationState.foldProjection";
+import { type LangyConversationStateData } from "@langwatch/langy";
 import type { LangyConversationTurnData } from "@langwatch/langy";
-import type { LangyMessageProjectionRecord } from "./pipelines/langy-conversation-processing/projections/langyMessageOperational.mapProjection";
+import type { LangyMessageProjectionRecord } from "@langwatch/langy";
 import type { LangyAnalyticsEventProjectionRecord } from "./pipelines/langy-conversation-processing/projections/langyAnalyticsEvent.mapProjection";
 import type { LangyTitleGenerator } from "../app-layer/langy/langy-title-generation.service";
 import type { LangyWorkerPort } from "../app-layer/langy/langyWorker";
