+++ langwatch/src/server/event-sourcing/pipelines/langy-conversation-processing/projections/index.ts
@@ -1,14 +1,12 @@
-export type {
-  LangyConversationState,
-  LangyConversationStateData,
-} from "./langyConversationState.foldProjection";
+// The spine's data type, fold, and the message record/map live in
+// @langwatch/langy (ADR-059) — import those from the package directly.
+export type { LangyConversationState } from "./langyConversationState.foldProjection";
 export { LangyConversationStateFoldProjection } from "./langyConversationState.foldProjection";
 // The turn document's data type, key helpers, and the fold itself live in
 // @langwatch/langy (ADR-059) — import those from the package directly.
 export type { LangyConversationTurn } from "./langyConversationTurn.foldProjection";
 export { LangyConversationTurnFoldProjection } from "./langyConversationTurn.foldProjection";
 export { LangyMessageOperationalMapProjection } from "./langyMessageOperational.mapProjection";
-export type { LangyMessageProjectionRecord } from "./langyMessageOperational.mapProjection";
 export {
   LangyAnalyticsEventMapProjection,
   type LangyAnalyticsEventProjectionRecord,
